refactor: rename api layer to web
