--- src/main/java/ru/jerael/booktracker/backend/api/dto/book/BookCreationRequest.java
@@ -1,20 +1,0 @@
-package ru.jerael.booktracker.backend.api.dto.book;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotEmpty;
-import jakarta.validation.constraints.Size;
-import ru.jerael.booktracker.backend.domain.constant.BookRules;
-import java.util.Set;
-
-public record BookCreationRequest(
-    @NotBlank
-    @Size(max = BookRules.TITLE_MAX_LENGTH)
-    String title,
-
-    @NotBlank
-    @Size(max = BookRules.AUTHOR_MAX_LENGTH)
-    String author,
-
-    @NotEmpty
-    Set<Integer> genreIds
-) {}
