--- src/main/java/ru/jerael/booktracker/backend/api/dto/genre/GenreResponse.java
@@ -1,3 +1,0 @@
-package ru.jerael.booktracker.backend.api.dto.genre;
-
-public record GenreResponse(Integer id, String name) {}
