+++ src/main/java/ru/jerael/booktracker/backend/web/dto/book/BookResponse.java
@@ -1,0 +1,17 @@
+package ru.jerael.booktracker.backend.web.dto.book;
+
+import jakarta.annotation.Nullable;
+import ru.jerael.booktracker.backend.web.dto.genre.GenreResponse;
+import java.time.Instant;
+import java.util.Set;
+import java.util.UUID;
+
+public record BookResponse(
+    UUID id,
+    String title,
+    String author,
+    @Nullable String coverUrl,
+    String status,
+    Instant createdAt,
+    Set<GenreResponse> genres
+) {}
