+++ src/main/java/ru/jerael/booktracker/backend/web/dto/genre/GenreResponse.java
@@ -1,0 +1,3 @@
+package ru.jerael.booktracker.backend.web.dto.genre;
+
+public record GenreResponse(Integer id, String name) {}
