--- src/test/java/ru/jerael/booktracker/backend/api/controller/GenreControllerTest.java
@@ -1,63 +1,0 @@
-package ru.jerael.booktracker.backend.api.controller;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.resttestclient.autoconfigure.AutoConfigureRestTestClient;
-import org.springframework.boot.webmvc.test.autoconfigure.WebMvcTest;
-import org.springframework.context.annotation.Import;
-import org.springframework.core.ParameterizedTypeReference;
-import org.springframework.test.context.bean.override.mockito.MockitoBean;
-import org.springframework.test.web.servlet.client.RestTestClient;
-import ru.jerael.booktracker.backend.api.config.ApiProperties;
-import ru.jerael.booktracker.backend.api.dto.genre.GenreResponse;
-import ru.jerael.booktracker.backend.api.exception.handler.GlobalExceptionHandler;
-import ru.jerael.booktracker.backend.api.mapper.GenreApiMapper;
-import ru.jerael.booktracker.backend.domain.exception.factory.GenreExceptionFactory;
-import ru.jerael.booktracker.backend.domain.model.Genre;
-import ru.jerael.booktracker.backend.domain.usecase.genre.GetGenreByIdUseCase;
-import ru.jerael.booktracker.backend.domain.usecase.genre.GetGenresUseCase;
-import java.util.Set;
-import static org.mockito.Mockito.when;
-
-@WebMvcTest(GenreController.class)
-@Import({GlobalExceptionHandler.class, GenreApiMapper.class})
-@AutoConfigureRestTestClient
-class GenreControllerTest {
-
-    @Autowired
-    private RestTestClient restTestClient;
-
-    @MockitoBean
-    private ApiProperties apiProperties;
-
-    @MockitoBean
-    private GetGenresUseCase getGenresUseCase;
-
-    @MockitoBean
-    private GetGenreByIdUseCase getGenreByIdUseCase;
-
-    @Test
-    void getAll_ShouldReturnSetOfGenres() {
-        Genre genre = new Genre(1, "adventure");
-        GenreResponse genreResponse = new GenreResponse(1, "adventure");
-        when(getGenresUseCase.execute()).thenReturn(Set.of(genre));
-
-        restTestClient.get().uri("/api/v1/genres")
-            .exchange()
-            .expectStatus().isOk()
-            .expectBody(new ParameterizedTypeReference<Set<GenreResponse>>() {})
-            .isEqualTo(Set.of(genreResponse));
-    }
-
-    @Test
-    void getById_WhenExceptionThrown_ShouldReturnNotFound() {
-        Integer genreId = 5555;
-        when(getGenreByIdUseCase.execute(genreId)).thenThrow(GenreExceptionFactory.notFound(genreId));
-
-        restTestClient.get().uri("/api/v1/genres/" + genreId)
-            .exchange()
-            .expectStatus().isNotFound()
-            .expectBody()
-            .jsonPath("$.detail").isEqualTo("Genre not found with id: " + genreId);
-    }
-}
