+++ src/test/resources/application.yaml
@@ -13,7 +13,7 @@ spring:
       max-file-size: 10MB
 
 app:
-  api:
+  web:
     max-file-size: ${spring.servlet.multipart.max-file-size}
   minio:
     url: http://localhost:9000
